+++ CSExeCOMServer/Program.cs
@@ -15,27 +15,19 @@
 * WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
 \***************************************************************************/
 
-#region Using directives
-using System;
-using System.Diagnostics;
-using System.Runtime.InteropServices;
-
-using ComTypes = System.Runtime.InteropServices.ComTypes;
 using System.IO;
-#endregion
 
 namespace CSExeCOMServerTest
 {
     class Program
     {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
+        // The main entry point for the application.
         static void Main(string[] args)
         {
-            // Comment below lines if Skype for Business is installed in the system.
+            // Comment below lines if the Unified Collaborations type library is already registered in the system.
+            string typeLibName = "UCCollaborationLib.tlb";
             string currentDir = Directory.GetCurrentDirectory();
-            string typeLibPath = $"{currentDir}\\lync4.tlb";
+            string typeLibPath = $"{currentDir}\\{typeLibName}";
             TypeLib.Register(typeLibPath);
            
             CSExeCOMServer.ExeCOMServer.Instance.OnCOMReady += new CSExeCOMServer.ExeCOMServer.OnCOMHosted(OnCOMReady);
@@ -48,108 +40,6 @@ namespace CSExeCOMServerTest
         static void OnCOMReady()
         {
             OutlookPresenceProvider.PresenceProvider.Started();
-        }
-    }
-
-    internal static class TypeLib
-    {
-        public static void Register(string tlbPath)
-        {
-            Trace.WriteLine($"Registering type library:");
-            Trace.Indent();
-            Trace.WriteLine(tlbPath);
-            Trace.Unindent();
-            ComTypes.ITypeLib typeLib;
-            int hr = OleAut32.LoadTypeLibEx(tlbPath, OleAut32.REGKIND.REGKIND_NONE, out typeLib);
-            if (hr < 0)
-            {
-                Trace.WriteLine($"Registering type library failed: 0x{hr:x}");
-                return;
-            }
-            hr = OleAut32.RegisterTypeLibForUser(typeLib, tlbPath, string.Empty);
-            if (hr < 0)
-            {
-                Trace.WriteLine($"Registering type library failed: 0x{hr:x}");
-                return;
-            }
-            Trace.WriteLine($"Registering type library succeeded.");
-        }
-
-        public static void Unregister(string tlbPath)
-        {
-            Trace.WriteLine($"Unregistering type library:");
-            Trace.Indent();
-            Trace.WriteLine(tlbPath);
-            Trace.Unindent();
-
-            ComTypes.ITypeLib typeLib;
-            int hr = OleAut32.LoadTypeLibEx(tlbPath, OleAut32.REGKIND.REGKIND_NONE, out typeLib);
-            if (hr < 0)
-            {
-                Trace.WriteLine($"Unregistering type library failed: 0x{hr:x}");
-                return;
-            }
-
-            IntPtr attrPtr = IntPtr.Zero;
-            try
-            {
-                typeLib.GetLibAttr(out attrPtr);
-                if (attrPtr != IntPtr.Zero)
-                {
-                    ComTypes.TYPELIBATTR attr = Marshal.PtrToStructure<ComTypes.TYPELIBATTR>(attrPtr);
-                    hr = OleAut32.UnRegisterTypeLib(ref attr.guid, attr.wMajorVerNum, attr.wMinorVerNum, attr.lcid, attr.syskind);
-                    if (hr < 0)
-                    {
-                        Trace.WriteLine($"Unregistering type library failed: 0x{hr:x}");
-                    }
-                }
-            }
-            finally
-            {
-                if (attrPtr != IntPtr.Zero)
-                {
-                    typeLib.ReleaseTLibAttr(attrPtr);
-                }
-            }
-        }
-
-        private class OleAut32
-        {
-            // https://docs.microsoft.com/windows/api/oleauto/ne-oleauto-regkind
-            public enum REGKIND
-            {
-                REGKIND_DEFAULT = 0,
-                REGKIND_REGISTER = 1,
-                REGKIND_NONE = 2
-            }
-
-            // https://docs.microsoft.com/windows/api/oleauto/nf-oleauto-loadtypelibex
-            [DllImport(nameof(OleAut32), CharSet = CharSet.Unicode, ExactSpelling = true)]
-            public static extern int LoadTypeLibEx(
-                [In, MarshalAs(UnmanagedType.LPWStr)] string fileName,
-                REGKIND regKind,
-                out ComTypes.ITypeLib typeLib);
-
-            // https://docs.microsoft.com/windows/api/oleauto/nf-oleauto-unregistertypelib
-            [DllImport(nameof(OleAut32))]
-            public static extern int UnRegisterTypeLib(
-                ref Guid id,
-                short majorVersion,
-                short minorVersion,
-                int lcid,
-                ComTypes.SYSKIND sysKind);
-
-            [DllImport(nameof(OleAut32), CharSet = CharSet.Unicode, ExactSpelling = true)]
-            public static extern int RegisterTypeLib(
-                 [In] ComTypes.ITypeLib typeLib,
-                 [In, MarshalAs(UnmanagedType.LPWStr)] string fileName,
-                 [In, MarshalAs(UnmanagedType.LPWStr)] string helpDir);
-
-            [DllImport(nameof(OleAut32), CharSet = CharSet.Unicode, ExactSpelling = true)]
-            public static extern int RegisterTypeLibForUser(
-                 [In] ComTypes.ITypeLib typeLib,
-                 [In, MarshalAs(UnmanagedType.LPWStr)] string fileName,
-                 [In, MarshalAs(UnmanagedType.LPWStr)] string helpDir);
         }
     }
 }
