+++ CSExeCOMServer/config.json
@@ -1,0 +1,4 @@
+{
+  "MattermostServerURL": "",
+  "MattermostSecret": ""
+}
