+++ DLL/UCCollaborationLib.IDL
