+++ OutOfProcessCOMBase/ExeCOMServer.cs
@@ -1,7 +1,5 @@
 ﻿#region Using directives
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Runtime.InteropServices;
 #endregion
