+++ OutOfProcessCOMBase/NativeMethod.cs
@@ -1,8 +1,6 @@
 ﻿
 #region Using directives
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Runtime.InteropServices;
 #endregion
